refactor: review & improve perfomance
