+++ src/core/async/core/clear.spec.js
