+++ src/core/async/core/index.ts
@@ -1,0 +1,19 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+/**
+ * [[include:core/async/core/README.md]]
+ * @packageDocumentation
+ */
+
+export { default } from 'core/async/core/api';
+
+export * from 'core/async/core/const';
+export * from 'core/async/core/helpers';
+
+export * from 'core/async/interface';
