+++ src/core/async/core/label.spec.js
