+++ src/core/async/proxy/index.ts
@@ -1,0 +1,18 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+/**
+ * [[include:core/async/proxy/README.md]]
+ * @packageDeclaration
+ */
+
+export * from 'core/async/core';
+
+export { default } from 'core/async/proxy/iterable';
+
+export * from 'core/async/interface';
