+++ src/core/async/proxy/proxy.spec.js
