+++ src/core/async/timers/index.ts
@@ -1,0 +1,18 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+/**
+ * [[include:core/async/timers/README.md]]
+ * @packageDocumentation
+ */
+
+export * from 'core/async/proxy';
+
+export { default } from 'core/async/timers/promise';
+
+export * from 'core/async/interface';
