+++ src/core/async/timers/interface.ts
@@ -7,7 +7,7 @@
  */
 
 import type Async from 'core/async';
-import type { IdObject, ProxyCb, AsyncOptions, AsyncCbOptions } from 'core/async/modules/base';
+import type { IdObject, ProxyCb, AsyncOptions, AsyncCbOptions } from 'core/async/core';
 
 export type TimerId = number | IdObject;
 
@@ -20,14 +20,14 @@ export interface AsyncWaitOptions extends AsyncOptions {
 
 export interface AsyncIdleOptions extends AsyncOptions {
 	/**
-	 * Timeout value for the native requestIdleCallback function
+	 * Timeout value for the native `requestIdleCallback` function
 	 */
 	timeout?: number;
 }
 
 export interface AsyncRequestIdleCallbackOptions<CTX extends object = Async> extends AsyncCbOptions<CTX> {
 	/**
-	 * Timeout value for the native requestIdleCallback function
+	 * Timeout value for the native `requestIdleCallback` function
 	 */
 	timeout?: number;
 }
@@ -36,3 +36,13 @@ export type IdleCb<
 	R = unknown,
 	CTX extends object = Async
 > = ProxyCb<IdleDeadline, R, CTX>;
+
+export interface AsyncRequestAnimationFrameOptions<CTX extends object = Async> extends AsyncCbOptions<CTX> {
+	element?: Element;
+}
+
+export interface AsyncAnimationFrameOptions extends AsyncOptions {
+	element?: Element;
+}
+
+export type AnimationFrameCb<R = unknown, CTX extends object = Async> = ProxyCb<number, R, CTX>;
