+++ src/core/async/timers/spec.js
