+++ src/core/async/wrappers/consts.ts
@@ -6,7 +6,7 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
-import type { EmitLikeEvents, AsyncOptions } from 'core/async/modules/events';
+import type { EmitLikeEvents, AsyncOptions } from 'core/async/events';
 import type { DataProviderMethodsToReplace } from 'core/async/interface';
 
 export const emitLikeEvents: EmitLikeEvents[] = [
